requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
Pillow~=9.0.1
torchmetrics~=0.9.1
wandb~=0.12.17